Do not expose explicit collection refreshes in the Account Screen
The PeopleSync app automatically synchronizes newly found address books,
and a sync also includes an implicit collection refresh.

This means, having a separate collection refresh action (which can
currently be triggered via the swipe down gesture and a separate floating
action button), is confusing.
If a new address book is found, it is shown as selected, but the actual
account would only be created once a sync has been initiated.

So, do not expose the collection refresh action in the Account Screen and
instead always use a full sync.

Only keep a single floating action button which triggers a full sync.

Trigger a sync + collection refresh when the account screen is opened.