Syncer: Trigger periodic collection detection
Whenever we periodically sync our address books, also trigger a
collection detection to find new addressbooks and delete stale ones.

As the Syncer is run as a background task, it's better to run
the collection detection/refresh synchronously before proceeding with
the address book updates. The list of collections will be up-to-date and
won't trigger any sync errors if a collection has been removed.

As such a refresh can fail, catch any errors and log them accordingly as
otherwise the actual address book syncs are not attempted.