Setup app to support managed configurations
This requires at least Android 5.0, which is fine, since this is already
our baseline. We no longer support Android 4.4.

This default_sync_interval setting is only used when creating new
accounts. Changing this value will not update exising accounts.
The interval is specified in seconds. The min interval is 900s i.e.
15min, as Android doesn't allow shorter sync intervalls.
The default interval is 14400s, i.e. 4h.

https://developer.android.com/work/managed-configurations