intro screen: Remove manufacturer warning from battery optimization screen
Instead add a nice background graphic for the regular sync intervals
setting.

We currently don't have a website providing more information how to deal
with overly aggressive battery optimization so hide the information in
the battery optimization page, including the "More information"
button which would only trigger a 404 when clicked.