Update version information for 4.4-0 release
So far our versioning scheme was based on the upstream version code,
shift it one to the left and increase the last digit whenever we make a
new release.

versionCode is an Int, the max value can only be 2^31-1.
With 2.6.6 this approach still worked but we were already pretty close
to that limit (2 060 600 000 < 2 147 483 647).
With 3.x we would exceed that limit, so the old versioning strategy no
longer works.

We now also need to tell the linter not to complain about that and use a
different versioning scheme, as we can't roll back those changes since
the 2.6.6-0 release was already published in the Play Store.

New versioning scheme:
2070 04 04 00
   |  |  |  |
   |  |  |  + PeopleSync version (monotonic counter for a given branch)
   |  |  |    Is reset to 0 for new branches
   |  |  + upstream minor version (based on branch name)
   |  + upstream major version (based on branch name)
   + fixed prefix