Trigger account migrations on startup
When upgrading from older versions, trigger a one-time migration of the
account settings to ensure the periodic sync workers are re-enabled
accordingly. Due to the WorkManager related refactorings, the old class
names are no longer available, leading to

2025-05-22 22:38:38.458 29143-29194 WM-WorkerFactory        com.messageconcept.peoplesyncclient  E  Invalid class: com.messageconcept.peoplesyncclient.syncadapter.PeriodicSyncWorker
                                                                                                    java.lang.ClassNotFoundException: com.messageconcept.peoplesyncclient.syncadapter.PeriodicSyncWorker
...
2025-05-22 22:38:38.459 29143-29194 WM-WorkerWrapper        com.messageconcept.peoplesyncclient  E  Could not create Worker com.messageconcept.peoplesyncclient.syncadapter.PeriodicSyncWorker

By triggering the account migrations, the periodic sync worker is
re-enabled with the new, correct class name.